A series of fixes that allow for pip-18.1 compatibility and improved cpython compatibility.
Specifically msgpack needs to unpack from a Py2kBuffer as a work around
on py<2.7.6 and from a bytearray directly on py>=2.7.7.  The struct
changes permit that.

It also needs to be able to concatonate a memoryview to a bytearray.

MemoryView also needs to throw TypeError if you attempt to take a
memorview of an ambigous type like a unicode to trigger the py2.7 work
around of passing it through buffer(Py2kBuffer) to set a default
utf-16-be encoding.