A series of patches to allow jython to work with msgpack without workarounds in upstream msgpack. This also allows pip-18.1 compatitiblity by way of dependency when combined with the Lib/zlib.py GZIP_TRAILER fix in decompress.
* src/org/python/core/Py2kBuffer.java

This is fairly cosmetic in that it moves __str__() functionality to
toString() and fixes the intended __repr__() customization that matches
cython and was broken in the 11 year old __repr__() derived object
recursion overflow workaround.

* src/org/python/core/PyMemoryView.java

This actually restricts and removes problematic functionality to make it
behave like cython which rejects attempts to take a memoryview of a
unicode object with a TypeError exception.  This is important for
pip/msgpack compatibility due to the fact that TypeError triggers use of
buffer to wrap the bytes in utf-16-be encoding rather than attempting to
strip it off entirely and potentially throw a unicode decode error.

* src/org/python/core/PyByteArray.java

This allows bytearray objects to __iadd__ PyMemoryView objects.
b=bytearray('asdf')
b+=memoryview('jkl;')

This is used by msgpack and supported by cython.

This also switches toString to create a string appropriate for java and
moves __repr__ functionality into bytearray___repr for cleaner repr
support and using the (String) value in struct.unpack methods.

* src/org/python/modules/struct.java

This set of changes allows struct to unpack from Py2kBuffer aka buffer
which is support by cython 2.7.x and used as a workaround for cython <=
2.7.6 which does not support unpacking from a bytearray.

It also adds support for unpacking from a bytearray to provide
compatibility without the workaround in msgpack.