src/templates/gderived-defs changes to prevent infinite recursion.
Ininite recursion is possible if the exposed method has the same
underlying method name as what it was exposed by and it is overriden in
any ...Derived.java class, but not it's children.

The mechanics behind this are a little weird, but it appears TYPE.lookup
search ast exposed method wrapped objects that call the underlying
method, so the lookup doesn't match the actual object method.

This should mostly not happen, but this patch should prevent it in any
case.  It might be worth some performance tests to see if it has any
side effects, but it should be a cheap test due to lookup cacheing.