This allows java String not already cast as a PyType to return an 8 bit clean PyString when possible instead of holding it to the higher bar of 7 bit clean by passing a PyUnicode to the upstream builtin.str.
The 8 bit permissive behavior is restricted to __str__ called on
java.lang.String object specifically, and reflects both the expected
return value from str and the ambiguity between byte values and string
in py2.7.

This seemed like a sensible change but was split out of the str / repr
recursion patch set as it is a behavior change not directly needed to
achieve the goals of the other patch set.